# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComposeBastion API",
    "version": "1.0.7-rc.1",
    "description": "ComposeBastion API contract. Stable JSON endpoints are available under /api/v1 while legacy /api routes remain compatible in V1."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cb_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": [
              "string",
              "null"
            ]
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "JobProgressStep": {
        "type": "object",
        "required": [
          "id",
          "label",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ]
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "OperationJob": {
        "type": "object",
        "required": [
          "id",
          "correlationId",
          "type",
          "status",
          "hostId",
          "payload",
          "result",
          "progress",
          "error",
          "createdBy",
          "createdAt",
          "updatedAt",
          "startedAt",
          "completedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "correlationId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "canceled"
            ]
          },
          "hostId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          },
          "result": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobProgressStep"
            }
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Backup": {
        "type": "object",
        "required": [
          "id",
          "hostId",
          "kind",
          "fileName",
          "status",
          "createdAt",
          "metadata"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "hostId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "volume",
              "host_path"
            ]
          },
          "volumeName": {
            "type": [
              "string",
              "null"
            ]
          },
          "sourcePath": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetVolumeName": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileName": {
            "type": "string"
          },
          "sizeBytes": {
            "type": [
              "number",
              "null"
            ]
          },
          "checksum": {
            "type": [
              "string",
              "null"
            ]
          },
          "backupTargetId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "remoteObjectKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "encryption": {
            "type": "string",
            "enum": [
              "none",
              "app_secret"
            ]
          },
          "encryptionKeyId": {
            "type": [
              "string",
              "null"
            ]
          },
          "encryptionKeyFingerprint": {
            "type": [
              "string",
              "null"
            ]
          },
          "verifiedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastDrillAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastDrillStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "completed",
              "failed",
              null
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "BackupsResponse": {
        "type": "object",
        "required": [
          "backups"
        ],
        "additionalProperties": false,
        "properties": {
          "backups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Backup"
            }
          }
        }
      },
...
      "BackupJobResponse": {
        "type": "object",
        "required": [
          "backup",
          "job"
        ],
        "additionalProperties": false,
        "properties": {
          "backup": {
            "$ref": "#/components/schemas/Backup"
          },
          "job": {
            "$ref": "#/components/schemas/OperationJob"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/backups": {
      "get": {
        "summary": "List backups",
        "tags": [
          "Backups"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a volume backup",
        "tags": [
          "Backups"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backups/{id}/download": {
      "get": {
        "summary": "Download a backup archive",
        "tags": [
          "Backups"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "File download stream"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}